pyproject.toml
[project]
license = { text = "Apache-2.0" }
authors = [
  { name = "Cloud-Dog, Viewdeck Engineering Limited" }
]
name = "cloud-dog-idam"
version = "0.3.1"
description = "PS-70 Identity & Access Management for Cloud-Dog services"
requires-python = ">=3.10"
dependencies = [
    "pyjwt[crypto]>=2.8",
    "argon2-cffi>=23.0",
    "cryptography>=42.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pyotp>=2.9",
]

[project.optional-dependencies]
config = ["cloud-dog-config>=0.2.0"]
logging = ["cloud-dog-logging>=0.2.0"]
oidc = ["httpx>=0.27"]
ldap = ["python-ldap>=3.4", "ldap3>=2.9"]
saml = ["python3-saml>=1.16"]
fastapi = ["fastapi>=0.110"]
db = ["sqlalchemy>=2.0", "alembic>=1.13"]
pam = ["python-pam>=2.0.2"]
mfa = ["pyotp>=2.9"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "httpx>=0.27",
    "fastapi>=0.110",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "ldap3>=2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
"Source" = "https://github.com/cloud-dog-ai/platform-idam"
